requirements.txt
alembic==1.7.7
autopep8==1.6.0
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
SQLAlchemy==1.4.35
toml==0.10.2
Werkzeug==2.1.1
zipp==3.8.0
